backend/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2-binary==2.9.5
pycparser==2.21
pytz==2022.6
sqlparse==0.4.3
